* Zwischenstand vor systematischen Commits

+++ db/app_migrations/003_create_graphics.rb
@@ -2,18 +2,18 @@ class CreateGraphics < Sequel::Migration
   def up
     create_table :graphics do
       primary_key :id
-      # foreign_key :user_id, :users # versions are linked to users
 
       String :title, size: 256, null: false
-      # String :description, text: true # variants should specify content
+      String :description, text: true
       DateTime :created_at
 
-      full_text_index :title
-      # full_text_index :description
+      full_text_index [:title, :description]
     end
   end
 
   def down
-    drop_table :graphics
+    if @db.table_exists?(:graphics)
+      drop_table :graphics
+    end
   end
 end
